#8.14 Bugfix

+++ src/controllers/videoController.js
@@ -16,14 +16,23 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   return res.render('edit', { pageTitle: `Edit: ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   // 위에 findById 대신에 exists 쓰는건, 굳이 video obj가 필요 없고
@@ -33,6 +42,10 @@ export const postEdit = async (req, res) => {
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found' });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -41,6 +54,7 @@ export const postEdit = async (req, res) => {
   });
   return res.redirect(`/videos/${id}`);
 };
+
 export const getUpload = (req, res) => {
   console.log(req.session);
   return res.render('upload', { pageTitle: 'Upload Video' });
@@ -73,6 +87,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.render('404', { pageTitle: 'Video not found' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
   await Video.findByIdAndDelete(id);
   // delete video
   return res.redirect('/');
